refactor: minimize the public API surface
