+++ client/JetStream/API.hs
@@ -5,55 +5,26 @@ module JetStream.API
   , consumers
   , publisher
   , messages
-  , accountInfo
+  , management
   , JetStreamRequestOption
   , withRequestTimeout
-  , Sequence
-  , sequenceFromWord64
-  , sequenceToWord64
-  , AccountAPIStats
-  , accountAPILevel
-  , accountAPITotal
-  , accountAPIErrors
-  , accountAPIInflight
-  , AccountInfo
-  , accountInfoTier
-  , accountInfoDomain
-  , accountInfoAPI
-  , accountInfoTiers
-  , AccountLimits
-  , accountLimitsMaxMemory
-  , accountLimitsMaxStorage
-  , accountLimitsMaxStreams
-  , accountLimitsMaxConsumers
-  , accountLimitsMaxAckPending
-  , accountLimitsMemoryMaxStreamBytes
-  , accountLimitsStorageMaxStreamBytes
-  , accountLimitsMaxBytesRequired
-  , AccountTier
-  , accountTierMemory
-  , accountTierStorage
-  , accountTierReservedMemory
-  , accountTierReservedStorage
-  , accountTierStreams
-  , accountTierConsumers
-  , accountTierLimits
   , JetStreamApiError
   , apiErrorCode
   , apiErrorCodeDetail
   , apiErrorDescription
   , JetStreamError (..)
-  , StreamName
-  , ConsumerName
-  , Subject
-  , Payload
-  , module JetStream.Consumer.API
-  , module JetStream.Message.API
-  , module JetStream.Publish.API
-  , module JetStream.Stream.API
+  , module JetStream.API.Consumer
+  , module JetStream.API.Management
+  , module JetStream.API.Message
+  , module JetStream.API.Publish
+  , module JetStream.API.Stream
   ) where
 
-import           JetStream.Consumer.API
+import           JetStream.API.Consumer
+import           JetStream.API.Management
+import           JetStream.API.Message
+import           JetStream.API.Publish
+import           JetStream.API.Stream
 import           JetStream.Error
     ( JetStreamApiError
     , JetStreamError (..)
@@ -61,29 +32,15 @@ import           JetStream.Error
     , apiErrorCodeDetail
     , apiErrorDescription
     )
-import           JetStream.Message.API
 import           JetStream.Options
     ( JetStream
-    , accountInfo
     , consumers
+    , management
     , messages
     , publisher
     , streams
     )
-import           JetStream.Publish.API
-import           JetStream.Stream.API
 import           JetStream.Types
-    ( AccountAPIStats (..)
-    , AccountInfo (..)
-    , AccountLimits (..)
-    , AccountTier (..)
-    , ConsumerName
-    , JetStreamRequestOption
-    , Payload
-    , Sequence
-    , StreamName
-    , Subject
-    , sequenceFromWord64
-    , sequenceToWord64
+    ( JetStreamRequestOption
     , withRequestTimeout
     )
