+++ client/JetStream/API/Consumer.hs
@@ -1,0 +1,4 @@
+-- | JetStream consumer management contract.
+module JetStream.API.Consumer (module JetStream.Consumer.API) where
+
+import           JetStream.Consumer.API
