+++ client/JetStream/API/Message.hs
@@ -1,0 +1,4 @@
+-- | JetStream message consumption and acknowledgement contract.
+module JetStream.API.Message (module JetStream.Message.API) where
+
+import           JetStream.Message.API
