+++ client/JetStream/API/Publish.hs
@@ -1,0 +1,4 @@
+-- | JetStream publishing contract.
+module JetStream.API.Publish (module JetStream.Publish.API) where
+
+import           JetStream.Publish.API
