+++ client/JetStream/API/Stream.hs
@@ -1,0 +1,4 @@
+-- | JetStream stream management contract.
+module JetStream.API.Stream (module JetStream.Stream.API) where
+
+import           JetStream.Stream.API
