--- client/Nats/JetStream/Consumer.hs
@@ -1,4 +1,0 @@
--- | JetStream consumer management.
-module Nats.JetStream.Consumer (module JetStream.Consumer.API) where
-
-import           JetStream.Consumer.API
