--- client/Nats/JetStream/Error.hs
@@ -1,16 +1,0 @@
--- | Errors returned by JetStream operations.
-module Nats.JetStream.Error
-  ( JetStreamApiError
-  , apiErrorCode
-  , apiErrorCodeDetail
-  , apiErrorDescription
-  , JetStreamError (..)
-  ) where
-
-import           JetStream.Error
-    ( JetStreamApiError
-    , JetStreamError (..)
-    , apiErrorCode
-    , apiErrorCodeDetail
-    , apiErrorDescription
-    )
