--- client/Nats/JetStream/Message.hs
@@ -1,4 +1,0 @@
--- | JetStream message consumption and acknowledgement.
-module Nats.JetStream.Message (module JetStream.Message.API) where
-
-import           JetStream.Message.API
