--- client/Nats/JetStream/Publish.hs
@@ -1,4 +1,0 @@
--- | JetStream publishing.
-module Nats.JetStream.Publish (module JetStream.Publish.API) where
-
-import           JetStream.Publish.API
