--- client/Nats/JetStream/Stream.hs
@@ -1,4 +1,0 @@
--- | JetStream stream management.
-module Nats.JetStream.Stream (module JetStream.Stream.API) where
-
-import           JetStream.Stream.API
