--- client/Nats/Message.hs
@@ -1,24 +1,0 @@
--- | Core NATS messages.
-module Nats.Message
-  ( Message
-  , MsgView
-  , Subject
-  , Payload
-  , Headers
-  , subject
-  , replyTo
-  , payload
-  , headers
-  ) where
-
-import           API
-    ( Headers
-    , Message
-    , MsgView
-    , Payload
-    , Subject
-    , headers
-    , payload
-    , replyTo
-    , subject
-    )
